Reorganize types module hierarchy
