+++ packages/liveblocks-react/src/types/errors.ts
